LayoutTests/imported/w3c/web-platform-tests/fetch/security/dangling-markup-mitigation-data-url.tentative.sub-expected.txt



PASS <img id="dangling" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII=">
FAIL <img id="dangling" src="data:image/png;base64,&#10;iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII="> assert_equals: Height expected 1 but got 0
FAIL <img id="dangling" src="data:image/png;base64,i&#10;VBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII="> assert_equals: Height expected 1 but got 0
PASS <img id="dangling" src="data:image/svg+xml;utf8,\n      <svg width='1' height='1' xmlns='http://www.w3.org/2000/svg'>\n        <rect width='100%' height='100%' fill='rebeccapurple'/>\n        <rect x='10%' y='10%' width='80%' height='80%' fill='lightgreen'/>\n      </svg>">
PASS <iframe id="dangling"\n        src="data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png'>\n        ">\n     </iframe>
PASS <iframe id="dangling"\n        src="data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/green-256x256.png?&lt;'>\n        ">\n     </iframe>
PASS <iframe id="dangling"\n        src="data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&amp;amp;lt;'>\n        ">\n     </iframe>
PASS <iframe id="dangling"\n        src="data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/green-256x256.png?&amp;amp;%2310;&lt;'>\n        ">\n     </iframe>
FAIL <iframe id="dangling"\n        src="data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
FAIL <iframe id="dangling"\n        src="     data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
FAIL <iframe id="dangling"\n        src="\ndata:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
FAIL <iframe id="dangling"\n        src="&#10;data:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
FAIL <iframe id="dangling"\n        src="\tdata:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
FAIL <iframe id="dangling"\n        src="\rdata:text/html,\n            <img\n              onload='window.parent.postMessage(&quot;loaded&quot;, &quot;*&quot;);'\n              onerror='window.parent.postMessage(&quot;error&quot;, &quot;*&quot;);'\n              src='http://localhost:8800/images/gr&#10;een-256x256.png?&lt;'>\n        ">\n     </iframe> assert_equals: expected "error" but got "loaded"
